PromiseArtifactId should be based on BaseDeferredKey and the index within the artifacts map
Summary: I think we are almost there to fix the promise artifact APIs. One thing I noticed when debugging the deduping issue is that the promise artifact ID's "id" field, which is equivalent to its index in the promise artifact registry, differs between anon targets, even though the artifact might end up being the same. Similar to how we have now changed the PromiseArtifact's BaseDeferredKey to be the anon target key, let's also have the id be the index of the artifact declared in the mappings, which should always be unique per anon target.

Reviewed By: cjhopman

Differential Revision: D49741336

fbshipit-source-id: 7b50b438b87170aa9f7296462d228c3c0c06a89e