Fix BuildCountManager
Summary:
Fixes two bugs with BuildCount:

1. The min_count should be only the min count of targets built within a
command (whereas right now it's the min of all targets ever built).
2. The count for a target should only be increased on successful commands. This
is the buck1 behavior and makes more sense (a build count of <=1 is intended to
indicate that its okay for the build to be doing slurp-build like work and after a
failed build we may have some of that work left to do, we only know we finish it after
a successful build).

Refactors BuildCount and related things to be harder to misuse, they now are only ever
based on resolved target patterns. If we get a failure before parsing patterns,
the build count isn't meaningful anyway.

Reviewed By: krallin

Differential Revision: D46539132

fbshipit-source-id: f8a72434e05ebc8291114c3c9e853e4c03907e39