Ensure [unstripped] subtarget exists for apple_library()
Summary:
Ensure we consistently expose an `[unstripped]` subtarget for `(apple|cxx)_library()`, as would be expected by dependents.

Previously, it would only be exposed on the default output, so that `[shared][unstripped]` would not work. This made it impossible to depend on being able to get the unstripped version of binary.

Reviewed By: chatura-atapattu

Differential Revision: D49952631

fbshipit-source-id: 8e11401235d0a9f0e57d4e1e4c2f8c6ac63669c8