be able to construct target set of action query nodes
Summary:
Aquery is weird because it takes in ActionQueryNodes, which are constructed after running analysis. Users usually pass in target labels or subtarget labels to aquery, so we need this function to run analysis and populate the aquery delegate nodes cache.

Alternative implementation is to add a `get_action_nodes()` on `AnalysisResult` in BXL, but that seems a bit more annoying to use and doesn't really allow for users to use aquery in BXL as they would on the CLI, or like cquery/uquery in BXL (ex: `ctx.aquery().deps("//foo:bar")`). Maybe we can add this too, but this current setup seems better to me for now.

Reviewed By: JakobDegen

Differential Revision: D49545271

fbshipit-source-id: a6b2a50b1921b0b7b3427106018912c122dbedb0