simply opam setup to just one link (#442)
Summary:
Pull Request resolved: https://github.com/facebook/buck2/pull/442

i convinced myself that creating a link to the ocaml standard library directory as extracted from `ocamlopt.opt -config` was necessary and that we can't just assume `$OPAM_SWITCH/lib/ocaml` but have failed to prove that to myself today. together with ocaml-scripts issue [#2](https://github.com/facebook/ocaml-scripts/issues/2) and similar past questions from vsiles i'm motivated to attempt to remove it so that we create but one link into .opam.

Reviewed By: vsiles

Differential Revision: D50082445

fbshipit-source-id: e46f2515c8c436cc5291550f6453f200defe0521