Add cfg constructor implementation for PACKAGE modifiers without selects
Summary: Add PACKAGE modifiers to cfg constructor. PACKAGE modifiers are passed in as a dictionary of constraint setting to modifier specified via PACKAGE files. For now, modifiers are only specified as constraint value targets and not modifier select types, but this diff sets up the foundation to add support for modifier select types.

Differential Revision: D49620942

fbshipit-source-id: 0a392592fd14b1862f679ac20df9961966911316