Add `set_cfg_modifier` PACKAGE file function
Summary:
This diff adds an initial skeleton of `set_cfg_modifier` function to set PACKAGE modifiers for all targets covered under a PACKAGE. Currently, only constraint values as modifiers are supported. `set_cfg_modifier` accepts a constraint_setting and modifier and sets a modifier for that constraint setting.

The underlying implementation uses PACKAGE values to write the modifiers. The key for the PACKAGE value is `cfg_modifiers.<constraint setting target with "." replaced>`. The value is a dictionary of modifier and the name of the PACKAGE that it was written from.

Note: this differs from the RFC in that this uses `set_cfg_modifier` instead of `cfg_modifiers` function. This is like the `add_default_cfg` function from the original RFC v1. I think this is better approach now after talking to more partners because it makes codemodding much easier. The downside is that it can make debugging more difficult. Will update the RFC accordingly once I get some bandwidth.

Differential Revision: D49483308

fbshipit-source-id: 36faf9680840c34f4b6f1480588dd32a56995840