Canonicalize naming the extra msvc toolchain folder targets we create
Summary:
These were a little organically grown. To make it easier for refactors
and adding new versions, ~~bikeshed~~ rename them so the arch substring matches
the arch foldername inside the msvc tools folder.

In the process I noticed that for arm builds we were pointing the host toolchain
to the arm compiler. That is... wrong. We only support windows x64 hosts. Fixed that.

Reviewed By: 8Keep

Differential Revision: D50041367

fbshipit-source-id: 836d0bc9cecd5020edc67ec34aca493a3db114a5