Move protos into a namespace mod
Summary:
A lot of our protos have the same or similar names to rust types and
it can be confusing when you encounter them without qualifiers. This just moves
some of them to be in a "protos" mod so that they are used with that qualifier
(their normal qualifier would be quite verbose).

Reviewed By: JakobDegen

Differential Revision: D50154464

fbshipit-source-id: d8e5f4ac9650dc35b14d1465c716437b252ddc3c