make it possible to disable fallocate
Summary:
The diff was written partially in response to S351880.

Some of the ai tests can you use a ton of disk space, and most likely there was
a regression somewhere. However while debugging it I've noticed that my
buck-out was barely compressed

{P782527780}

despite btrfs being mounted with `compress-force=zstd:3`. We've seen problem
like that before, and usually it could be related to fallocate not being
disabled. This diff makes it possible to disable it, but hides it behind a
flag, since we'd probably want to roll it out carefully (if we want to roll it
out at all).

Reviewed By: arauter

Differential Revision: D47273708

fbshipit-source-id: 86f1e6636227b0b632df867f4781eb4f489a117f