buck2_error: Support downcasting
Summary:
Adds downcasting support to `buck2_error::Error`.

This has one notable limitation (demonstrated in a test): It is not in general possible to downcast to context that was added via `anyhow::Context`. That's unfortunately impossible to support, since anyhow just does not provide sufficient APIs to be able to extract the context from its errors in a useful way.

That being said, I don't think that has to be a huge deal. The plan for making use of this/rolling it out should be as follows:

 1. Find all the places where we do error downcasting and make them use `buck2_error::Error::downcast_ref`.
 2. Find all of the places in which we add context that we expect to be able to downcast to, and make sure those places use `buck2_error::Context`.

This is all the more reason that we should probably ban `anyhow::Context` and just use `buck2_error::Context`. Having to support everything simultaneously is just too hard.

Reviewed By: cjhopman

Differential Revision: D50213704

fbshipit-source-id: cad8170a58c23dc2527e25945435c19e0c082371