buck2_error: Use `anyhow` in `ErrorRoot`
Summary:
When I was testing the rest of this stack, I ran into a bug that turns out to be a little bit non-trivial to fix. In a number of places we currently downcast `anyhow::Error`. That already doesn't quite work naturally because `SharedError` interferes with that - however, that gets worked around by introducing yet more downcasting.

Introducing `buck2_error::Error` into error stacks interferes with this even more. The conversion from `buck2_error::Error` to `anyhow` involves wrapping in a `CrateAsStdError` wrapper, and anyhow won't be able to look past that wrapper, at least when downcasting to context (but also in other situations).

The fix for all of this is to properly support downcasting `buck2_error`, and then use that everywhere. However, that requires preserving anyhow errors in the error root. If we instead handle them as std errors, we lose some of the information about error downcasting.

Despite this change, we still require a more or less well-behaved interface on the `ErrorRoot` API, and make it as difficult as possible to stuff arbitrary garbage into the error root.

Reviewed By: stepancheg

Differential Revision: D50213702

fbshipit-source-id: 956bb6c08ea79a024b61f0e694415993c150d5b1