buck2_error: `ErrorRoot` in a separate file
Summary:
This is gaining complexity in the next few diffs.

More generally, the code in this crate is already really hard to read. The ratio of actual executable code to random type/trait definitions and impls that do nothing but forward to other impls is really low. Keeping things well encapsulated and keeping file sizes down helps mitigate this a bit

Reviewed By: cjhopman

Differential Revision: D50213719

fbshipit-source-id: db79d7416a68d7067c5db86d2fc99942665c3ab7