errors: Use `buck2_error::Context` in a couple places
Summary:
As mentioned earlier in the stack, if we intend to downcast to some context, using `anyhow::Context` might be wrong. There's only two types that I know of with which we do this, and this diff makes sure to use `buck2_error::Context` with those types.

This is obviously not really a desirable state of affairs. In the coming weeks I'm going to move this `ErrorReport` thing to be more general purpose and properly integrated with `buck2_error`, in which case the footgun around using the right context type can be made to go away.

Reviewed By: wendy728

Differential Revision: D50213699

fbshipit-source-id: 84c63a5d67b5ffc37f5ac5bf7077f2d74d810024