buck2_error: Be `SharedError` aware
Summary:
The goal is obviously to get rid of `SharedError`. However, doing that in one megadiff is not feasible. So instead, let's support it in `buck2_error`.

This diff finally brings `buck2_error` to a state where it works with everything - that is, when unsure, using `buck2_error` is always the right choice. That should mean that we can fearlessly migrate things that need migrating.

Reviewed By: cjhopman

Differential Revision: D50213713

fbshipit-source-id: 086e378495d84a75b6c5ecbeb41e0e0c1a8ddb4d