Fix PAR_MAIN_OVERRIDE when main function is set
Summary:
This is an attempt to fix the behavior of `PAR_MAIN_OVERRIDE` when a main_function is set instead of a main module.
For that, I am using a new environment variable named `PAR_MAIN_FUNCTION_OVERRIDE` to allow users to specify the function entrypoint in the decorator. If a main function is set, then it is up to the decorator to run (e.g., via `__par__.bootstrap.run_as_main`).

Reviewed By: zsol

Differential Revision: D50364859

fbshipit-source-id: 2e05d575b28cb55d529b77bc7b1e5e2a5a357657