Revert D49976387: Multisect successfully blamed "D49974422: [buck2][python] use anon targets for compilation" for test or build failures
Summary:
This diff is reverting D49976387
Depends on D50401636
D49974422: [buck2][python] use anon targets for compilation by BrandonTheBuilder has been identified to be causing the following test or build failures:

Tests affected:
- [registry/builder/fbpkg/tests:test-caf-builder-e2e - test_e2e (registry.builder.fbpkg.tests.test_caf_builder_e2e.CafBuilderE2ETestCase)](https://www.internalfb.com/intern/test/281475088848239/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3330598
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: lmvasquezg

Differential Revision: D50401645

fbshipit-source-id: cd3ee79dc76ac6f4698791e7a8ac7ef7cdd65ae7