add an option to dump all produced artifacts to a file
Summary:
Main motivation for this change is closer integration with [CAF](https://fb.workplace.com/groups/contentaddressedfs), which is a new way of storing fbpkg. Currently when building CAF package we have to [iterate over all artifacts produced by buck2](https://www.internalfb.com/code/fbsource/[53af288d0950]/fbcode/fbpkg/caf/src/upload_map/mod.rs?lines=226) in order to find what nees to be uploaded. On a good day it takes ~2-4 secs (tolerable, but not great), however it could spike to 20 secs or more, which significantly
increases build&packaging time.

This diff makes it possible to ask buck to print all artifacts to a file, which
CAF can later read and avoid traversing the directory.

I serialized the artifact data into json, mainly for simplicity and ease of debugging.
The structs are defined in a crate which I plan to depend on from CAF so that
deserilization works the same way. Lmk if it's not a good idea, or if you have
ideas on where to put these structs.

Reviewed By: krallin

Differential Revision: D49055340

fbshipit-source-id: f018460c88c49205fae191789942f4b2af033ce7