build report: Swap two names
Summary:
Maybe I'm missing something. But the thing that is now called `BuildReportEntry` is always associated with an unconfigured target. The thing that is now called `ConfiguredBuildReportEntry` is associated with a configured target. That's the version of things that makes sense to me.

These names were previously the other way around.

Reviewed By: wendy728

Differential Revision: D50260487

fbshipit-source-id: 6aed6745bde553466d5558dfc47a9afd04f97a4a