build: Configuration errors in the build report
Summary:
Well, not quite. Errors in producing a `ConfiguredTargetNode` actually get to us through analysis, not the codepath touched here, and so are in the build report already. The affected codepath is specifically for errors computing the target platform; for example, if it refers to a non-existant target like in the test.

This uses the field we added last diff. Making use of it requires introducing a `BuildEvent` wrapper around `ConfiguredBuildEvent` which can represent the error

Reviewed By: wendy728

Differential Revision: D50303572

fbshipit-source-id: f4bc7674acf5f93c6ce58cf7e03617418f3aa772