build: Report more errors to ingress
Summary:
Right now there are two ways that a build command can return.

The first of these is the case in which it produces a `BuildResponse`. In this case:

 1. The only possible errors that may have occured are action errors
 2. The command can produce additional build metadata, like for example a build report
 3. None of the reported errors show up in the "error messages" field in `scuba buck2_builds`

The second is the case in which it produces an `anyhow::Error` instead; in this case, the opposite of all the things above apply.

In order to make it possible to include non-action errors in the build report, we need to change 1. However, that would cause a regression in telemetry because of 3. This diff changes 3., ie it ensures that all error messages are reported via the invocation record to telemetry.

Note that for action errors, this does not mean that the entire action error shows up in the "error messages" field; instead, it displays only the "failed to build target" message that is shown at the end of the command.

Personally, my take is that these errors being missing from buck2_builds has always been a bug anyway.

Reviewed By: wendy728

Differential Revision: D50311160

fbshipit-source-id: 58b3ce8b9c3cfe14ba892a997338eb3bd0681b28