build report: Report errors
Summary:
Bit of a megadiff so I apologize for that.

This diff puts errors into the build report. The current implementation is the most naive one possible. It has two major limitations:

 1. No build report for anything other than action errors.
 2. Subpar error deduplication.

Mostly this diff is there to set up tests. Both of these limitations will be fixed in this stack.

Reviewed By: wendy728

Differential Revision: D50228649

fbshipit-source-id: 04a9539763f5b0bcc09c4e13f3c9842facc21353