Remove a bunch of TODOs
Summary:
There are a number of comments in the codebase right now that indicate that at some time we were intending to move a bunch of code that currently sits in the server into the client. It's been multiple years since those comments were added though, and that hasn't happened. I also don't see an obvious push for it at this time.

I'm putting this diff up under the claim that these comments no longer reflect the state of the world. The way things are right now does not seem obviously broken to me, and if we really did want to go forward with these changes, I'd prefer to see some new discussion about whether it's the right direction first.

I can drop this commit if people disagree

Reviewed By: wendy728

Differential Revision: D50311162

fbshipit-source-id: 20a8434aa01eb2edba82492b1f6c55301187bf40