build report: Inline error handling
Summary:
Instead of post processing some of the errors, do it all inline. This code has the unfortunate property of being a little bit clever and using a little too much iterators, but I've done my best.

This mostly concludes the refactoring. There resulting code is not really shorter, but it is a lot easier to follow, and importantly, to chagne. This is primarily thanks to it being more dataflow based, instead of being a jumbled mess of shared mutability.

Additionally, each of the fields in the build report are now handled in one place only, which is going to make follow ups a lot easier.

There's still a lot more to do here in terms of improving code quality, but for now, I'm going to focus on deliverables

Reviewed By: iguridi

Differential Revision: D50397284

fbshipit-source-id: 7e1ec61bfe17696360d1b9e56fb21ce81a59cabb