Don't forget about action errors
Summary:
What most of this stack has lead up to.

Instead of producing an action error event and then forgetting about the error, keep the action error around. We mark it as emitted, and use the late format mechanism to make sure we can still report the issue at the end of the program.

This should hypothetically not change any behavior. We'll see what CI has to say about that

Reviewed By: wendy728

Differential Revision: D50228652

fbshipit-source-id: 5a4a712cda0b679c3d8ffb2da1ead1cde898092e