build: Remove `BuildOwner`
Summary:
This was added more than a year ago, with the anticipation that it would grow more variants; however, that never happened. In the meantime, as I'm doing this refactoring, it's unclear to me which code is actually supposed to be generic over the build owner. That means its likely that whenever this does get a use case, it would need a bunch of changes anyway.

Just delete it for now. If someone ends up needing it, they can put it back

Reviewed By: iguridi

Differential Revision: D50397279

fbshipit-source-id: cdf9263576dc73617241eb7c9323686e92676104