build: Start refactoring `ResultReporter`
Summary:
The structure of this doesn't really make any sense right now. We collect the `BuildEvent` stream into `BuildTargetResult`, only to then piece that apart again by iterating through it, and recollecting the result into the build report and `ResultReporter`. Instead of doing all that, let's just refactor this to a pair of functions that goes from `BuildTargetResult` to `ResultReporter` and `BuildReport`

This diff starts with the `ResultReporter`. I'll do this in small diffs to keep things easy to review

Reviewed By: iguridi

Differential Revision: D50397305

fbshipit-source-id: be4ed218cdfbfb26a12e6605233e7c0066fc82b1