build: Errors field for unconfigured errors
Summary: Adds a field to the `BuildTargetResult` for errors that cannot be associated with a particular configured target. This diff just includes the code to handle these errors, it does not yet include the code to fill them out; left that for the next diff, since that's big enough as is.

Reviewed By: wendy728

Differential Revision: D50303573

fbshipit-source-id: 53cb50e139b6b7ced51ed030b4ddd2f635f09df6