build report: Report error duplication
Summary:
Adds information about error duplication to the build report. Like described in D50228655 , this is done via a deduplication index on errors which is identical when the errors are duplicates.

This diff does not actually deduplicate any errors in the build report; instead, we only report the duplication and leave it to consumers to actually act on that information. That is probably not the ideal long term choice; however, deciding which cases of error duplication can be safely omitted is not straightforward, and is not necessary for initially getting this feature off the ground.

The implementation is based on the `UniqueRootId` introduced to `buck2_error` in the previous diff. This works quite well already, with no need for any additional changes to the code base. The reason for that might be non-obvious; it's the result us having already replaced `SharedError` with `buck2_error::Error`. This means that every error is turned into a `buck2_error` at latest when it hits the end of a dice computation, which is already plenty error duplication information to do a reasonably good job.

The existing tests demonstrate that this works swimmingly - the error deduplication is correct on all of them.

It's worth considering how this will affect maintainability of the codebase. In practice, this currently couples our notion of error duplication to the dice graph in a potentially undesirable way. There is potential for this to become a refactoring hazard.

One significant way that we can mitigate this effect is by more reliably creating `buck2_error::Error` instances earlier, for example by consistently replacing `thiserror` with `buck2_error`, which ensures that a `buck2_error` is always created as soon as the error is.

At the same time though, I also don't see an alternative to this approach. Manually maintaining annotations of some kind that indicate error duplication is unlikely to scale.

Reviewed By: stepancheg

Differential Revision: D50397303

fbshipit-source-id: 3d734a7fc635ed15f4483e6f38d35216d1a3a50c