upload Action to CAS when uploading ActionResult to Action Cache
Summary:
We want to implement new security measures in the ActionCache, so that every platform can write results to the AC belonging only to its own platform.

We do this validation at write time: when a ActionResult is written to the AC, we check that its corresponding Action belongs to the same platform of the writer.

When Buck stores locally executed actions to the ActionCache, this validation would fail since Buck doesn't store locally executed Action to CAS.

In this diff we want to add Action writes to CAS

Reviewed By: stepancheg

Differential Revision: D50330425

fbshipit-source-id: 19f4c7a41ea1584b380af85e0d1516849f5bbc72