Fix error formatting
Summary:
Originally, action errors were always printed with `":#"`. This was fine, because they never had context, so it doesn't matter.

However, changes that landed yesterday accidentally moved the action error codepath to also include lots of other errors, which caused the breakage shown in the diff below this.

That's not hard to avoid

Reviewed By: shonaganuma

Differential Revision: D50506421

fbshipit-source-id: ba1c2feebe9c0a95c35ce6e8b8bfed1413d7d14b