Make eval parameter typed
Summary:
For documentation, typechecking, also less code.

This diff adds unsupported types to function signature to make an error that type is not supported. ("target_set with configured nodes are currently not supported").

I think this is overkill: function signature should make it clear that it is not supported. Preserving current behavior but proposing to remove it in the following diff D50483955.

Reviewed By: JakobDegen

Differential Revision: D50483958

fbshipit-source-id: 840b212cc41e12d23b32fc477b793a34081fd52c