Add platform prefix to stripped outputs for prebuilt libraries
Summary:
To prevent path conflicts under the `stripped` output of the binary with multi-arch Android somerge builds.

With non-prebuilt libraries, the unstripped library is created during somerge at a path with the platform prefix added, if there are multiple platforms.
With prebuilt libraries we are not creating a new artifact so it's necessary to specify a path instead of re-using the path of the unstripped library.

Reviewed By: smeenai

Differential Revision: D50480779

fbshipit-source-id: af9a101c551ddc464df838553b900a8d1555a9d6