Bundle only .app for dmg
Summary:
# Request For Comments
I'm a buck2 noob. Please help me shape this.

## Before

The current `packager` attribute passes on the `__unzipped_ipa_contents__` to the packaging script. That means we bundle up a lot more than we previously did. This leads to this experience when opening a dmg:

 {F1129506335}

This changes the behaviour to go from `--contents-dir` back to the `--app-bundle-path` as it existed in the original version of the script, pointing directly to the app bundle instead.

https://www.internalfb.com/code/fbsource/[df65d81c4468]/fbobjc/Tools/macosx/package-dmg.sh?lines=7

NOTE: There is still a discrapancy between the old buck1 approach and the `packager` command as the metadata name of the dmg will not match the application. We will handle this on the packager script level.

Reviewed By: d16r

Differential Revision: D50600912

fbshipit-source-id: 77f6b5aafee68c59f405f0594aa0219ec1e43634