Make MetadataMap sorted
Summary: Make `metadata` attribute sorted by using `SortedMap` instead of `OrderedMap`. Sorting adds a little bit of compute overhead, but a fixed order can improve caching. When  the ordering of metadata insertion is switched, it can still get cache hits on DICE. `metadata` is not visible in analysis, so no one should rely on this ordering in practice.

Reviewed By: stepancheg

Differential Revision: D50662546

fbshipit-source-id: a01431dbcdd3d4ba87687fd8622c9a0168829b5f