Disable eager write dispatch on macs
Summary: This mitigates the issue we are seeing where materialization seems to be stalling on android, where the real issue seems to be slow writes queued up that prevent materialization from taking place in the first place. This diff fixes this by disabling  writes from being eagerly dispatched on macs.

Reviewed By: christolliday

Differential Revision: D50723694

fbshipit-source-id: 453b24f989cba5ef90ac261621ee05e4178dddb3