client: Do error code downcasting a little differently
Summary:
What we do right now is pretty weird. For basically three identical scenarios, we have three instances of a `FailureExitCode` type, all of which map to the same exit code.

What this code actually wants to do is to recognize when an operation has failed because of io with "the user." So let's just make a marker type that has that meaning, and then downcast to that directly, instead of having a bunch of extra variants.

Reviewed By: stepancheg

Differential Revision: D50689243

fbshipit-source-id: e2c23b6eea3581e61346032ea7a7234087dd994a