errors: Report no category if we don't have one
Summary: The error report generation has so far has fallen back to "infra" if there was no category on the error. Instead of doing that, let's just not report a category at all. Users can fall back to infra on their own, but the fact that an error was uncategorized might be useful

Reviewed By: stepancheg

Differential Revision: D50689267

fbshipit-source-id: 942fca6c727ef62fc26001c6536a13143e9ce885