errors: Add `errors` field to invocation record
Summary:
I was originally going to just put the `ErrorReport` into the invocation record. However, that presents a bit of a challenge due to the nature of protobuf enums. On the one hand, we want the `ErrorCause` to be an actual enum, since we want to match on it or whatever within buck2. On the other hand, we don't want to have to deal with needing to keep backcompat so that users of the invocation record can have all the variants they might care about available. Also, JSON representation of the enum don't convert it to a string but keep it as an int, which makes it extra useless. Preprocessing the field into a string on the buck2 side makes a lot of things easier.

I don't expect that this is going to be the last field for which we want to do something of this sort.

Reviewed By: stepancheg

Differential Revision: D50689222

fbshipit-source-id: e9a411926f35748d01f83e4eb07c9d8c114ef007