client: Remove `ExitResult::failure`
Summary: This function is a footgun. Callers should have access to some kind of structured error information from which to inform exit code creation, and not be making up new generic failures. Both the one remaining use in this diff and the two uses in the previous diffs are effectively bugs. We can prevent these from being re-introduced

Reviewed By: stepancheg

Differential Revision: D50689226

fbshipit-source-id: 6d0df4a475f6c517a86ea58719e85ff0f4a4589b