errors: Put `create_error_report` where it belongs
Summary:
I said many diffs ago that I'd put this back, and now we can finally do that.

This doesn't need to be in buck2_common. All of the conversion of errors to `ErrorReport`s happens when commands end, and all of the command ending logic has access to `buck2_server_ctx`

Reviewed By: stepancheg

Differential Revision: D50689249

fbshipit-source-id: d922d3deb22b24bf732dcfa7908ef89dff971428