client: Delete legacy error categorization
Summary:
With our `ErrorReport` and `buck2_error::Error` powered error categorization fully on-par with what we did before, we can delete the old error categorization strategy.

The advantages of the new version are that it is a bit more robust and extensible, but also that it follows the typical data-flow much more closely - we no longer need to keep state in the invocation record to do error categorization, we can just directly look at the errors from the daemon.

Reviewed By: stepancheg

Differential Revision: D50689217

fbshipit-source-id: cfef0898bbfbf105cb06e02c8e1946f65f624fd7