Let the rule cope with a list
Summary:
The attrs for Haskell already dealt allowed a list of sources, but the analysis didn't handle it right. Fix that up.

Eventually we should probably require a list of sources, since I don't think its going to work if your filename and string name for the artifact don't match up.

Main reason is to allow more sensible Haskell targets in open source (internally we wrap them in a macro anyway).

Reviewed By: pepeiborra

Differential Revision: D50735764

fbshipit-source-id: 50ba4fbc8b2a3510383cb78c0e1f2e9168656db7