Convert ValueAsCommandLineLike to struct
Summary:
And implement `UnpackValue` for it.

So it could be used as native function argument, like in the following diff D50672044.

This is not done well, because we cannot for now give better `StarlarkTypeRepr` for `ValueAsCommandLineLike` than `Any`.

To fix if we should either:
* enumerate all implementations in `ValueAsCommandLineLike` (which is hard because types live in two crates)
* implement some starlark abstract type for command line arg like (also not easy)

When reviewing this diff, consider reading `buck2/app/buck2_build_api/src/interpreter/rule_defs/cmd_args/value_as.rs` first. The rest of the diff is mechanical change.

Reviewed By: JakobDegen

Differential Revision: D50671874

fbshipit-source-id: 79b85eefeb82406ce7f5d68f6e83c34f7450b7f7