Add additional zig releases to zig toolchain (#464)
Summary:
More recent version of zig are required to compile code on modern versions of macOS (at least 13.x, and probably 14.x as well), otherwise zig seems to be unaware of the macOS version and errors out.

I tested this on Linux and it still works. It gets _further_ on macOS, but runs in to more problems with the linker even after https://github.com/facebook/buck2/issues/462. I'll file a separate issue for that once this is accepted so that the error can be reproduced on the main branch.

Pull Request resolved: https://github.com/facebook/buck2/pull/464

Reviewed By: JakobDegen

Differential Revision: D50770987

Pulled By: ndmitchell

fbshipit-source-id: 76a8110edf0c31f1b8b631df05aa42934a151963