buck2_error: Change derive macro strategy
Summary:
The last part of this stack is to improve the ergonomics of setting a non-default error category or type.

The point of the `buck2_error::Error` derive macro is to customize the behavior of the error. That means we need to be able to reliably hijack the conversion from the error type to not just a `buck2_error::Error`, but to an `anyhow::Error` as well. In addition to that, we also want to avoid reimplementing thiserror.

Unfortunately, these goals are a bit at odds. Because we have no eager macro expansion, we cannot modify the code produced by thiserror. But without modifying that code, there would be no place in which to hijack the behavior of `anyhow::Error::new(ErrorType::Variant)`.

As a result, the API that we'll provide from the macro will change a bit. Instead of providing a `Display` and `StdError` impl the way `anyhow` does, we'll just provide impls of `Display`, `Into<anyhow::Error>`, and `Into<buck2_error::Error>`. This is not quite as good as `StdError`, but it should support all the common use cases.

The details of the implementation are well-explained inline

Reviewed By: stepancheg

Differential Revision: D50689241

fbshipit-source-id: a091a92ced55f49689ea0556af4ca02e36e6a844