Add option to sort what-materialized by size
Summary:
On big builds, output of what-materialized could be huge and when one wants to investigate what took so long, it is hard to understand directly which files took a lot of space. This diff adds the ability to sort the output of what-materialized by size.

TODO: I think it would be good to add another option (or replace this altogether) that gets the top-k by size as well.

Reviewed By: stepancheg

Differential Revision: D50733739

fbshipit-source-id: d5e1d445a65aae30388c9a2a7f94ce41548db79c