client: Fix "ctrl-c was pressed" bug
Summary:
I noticed today that whenever I ctrl-c'd buck, it's been printing "command failed: ctrl-c was pressed" to stderr. This is a bug that was accidentally introduced in the refactor at D50689224 .

We can fix it by just not having an error type for this situation. We don't need one anyway

Reviewed By: shonaganuma

Differential Revision: D50833346

fbshipit-source-id: 71dd1cd5b9b4ad7b2b5e31e8c20230abfa03c535