errors: Add back `ErrorCategory::USER`
Summary:
But this time, using a non-zero value for it in the protobuf spec. Zero is the default value, which we shouldn't use for actual data.

Not emitted yet, but support added to the "read" side. Needs to be landed first, subsequent diffs will start emitting this and then remove the old version

Reviewed By: wendy728

Differential Revision: D50828579

fbshipit-source-id: 5a39f9bd28e7d84be759264c70a0ee4ad9d7f012