macros: Add `$(source ...)` support
Summary:
Asked about this in buck2 dev at some point and feedback seemed positive. Not hard to implement, and mostly behaves as you'd expect.

Decisions in this diff were to allow directories (I don't see why not), and not allow targets. This second point is inconsistent with `attrs.source()`, but on the other hand we already sort of have `$(location )` for that

Reviewed By: wendy728

Differential Revision: D50781532

fbshipit-source-id: ca865e468307b3c8e0cec47a97329192f1f1fbb1