errors: Replace some `anyhow::Error::new` with `anyhow::Error::from`
Summary:
I'll first briefly explain what I want to do after this diff: I'd like to codemod all uses of `thiserror::Error` in `buck2/app` to instead use the `buck2_error::Error` derive. It was prievously unclear whether this would be necessary, but I now think it is and we should do it. Two reasons:

 1. stepancheg and I discussed earlier today that our current usage of `thiserror` possibly doesn't make all that much sense anyway. The benefit of having to declare the error in a place different from where it's instantiated is at least unclear. I'm not immediately proposing that we change anything here, but if we do want to change something in the future, replacing `thiserror` with `buck2_error` is at least the small step down that path. It allows us to start locking down the API that people can use and eg requiring that an error type be supplied (if we want to do that).
 2. I want to start reporting the source location of the error to telemetry. I was originally looking at accomplishing that without needing to do this, but that's unfortunately going to have its limitations. We can still get most of the way there, but using `buck2_error::Error` derives everywhere will more reliably get us more accurate source information, which seems useful and worthwhile.

This diff: Replaces a couple uses of `anyhow::Error::new` with `anyhow::Error::from`. `buck2_error::Error` derived errors don't (and can't) work with `anyhow::Error::new`, so we need to make this change before we can switch the derives over

Reviewed By: stepancheg

Differential Revision: D50829502

fbshipit-source-id: 5b55a1a437a2410619b070a90698527d0aca6407