Capture more instruction count metadata in build logs
Summary:
We use instruction counts for regression detection and monitoring, and recently observed some unexpectedly high instruction counts (e.g. https://fburl.com/gsheet/xyhk2nz6, D50575311). While the root cause is not known yet, one hypothesis is that this could be due to PMU multiplexing and scaling.

The necessary information is already available in RE and this diff adds it to buck2 logs. I used Thomas's diff D45236116 as a reference for the locations that need to be updated. Next steps are to read it from build logs and log it during regression detection to verify the hypothesis and understand if we should be using different thresholds when scaling is enabled. We don't need this ratio in scuba yet, but I'm happy to make the necessary changes if it's useful more broadly.

Reviewed By: krallin

Differential Revision: D50609936

fbshipit-source-id: 7ea20aba880653655d453cf938fc37bee4efb68f