Don't clean artifact if its deps fail to materialize
Summary:
There is a bug in the materializer where:
- Artifact A depends on Artifact B
- User runs a build which calls `ensure_materialized(A)`, which materializes B
- B changes (ex. rebase)
- User runs a second build where:
  - `ensure_materialized(A)` is called, A is already materialized, but B fails to materialize (ex. CAS download error)
  - A is cleaned, but not invalidated
  - Command fails because materialization fails
- User runs a third build where:
  - `ensure_materialized(A)` is called, this time B is successfully materialized but A is not re-materialized since it is still in a materialized state despite being cleaned.
  - The command succeeds but A does not exist in buck-out
- Subsequent builds can not materialize A without re-initializing materializer state (running buck clean)

This resolves this by simply not cleaning A if B fails to materialize. This leaves the materializer in a state where A is materialized but B is not, but this should be safe as long as we only assume A and it's dependencies are materialized if `ensure_materialized` does not return an error, which should already be the case.

In practice this is being hit most often by Android install commands because they materializes symlinks (native_libs_symlink and others) that point to thousands of dependent artifacts, so if CAS reliability is degraded at all, the probability of a single dependency failing to materialize is relatively high.

Reviewed By: scottcao

Differential Revision: D50811059

fbshipit-source-id: b2a8001c953984e1e49257da0e9fbd7488a2a4f5