Support multiple modifier selects in `set_cfg_modifier`
Summary: Currently, we only support one package modifier for a constraint setting at a time. However, with `modifier_select`, it's possible that multiple modifier_select's are specified and all of them need to be evaluated. This diff adds support for multiple modifiers for the same PACKAGE value in `set_cfg_modifier`. We do this by storing a list of `CfgModifierWithLocation` instead of a single one for every modifier constraint setting. When `cfg_modifiers` is called, we check what the modifier passed is. If it's a modifier select, we call `read_parent_package_value` and add the modifier select to an existing list and then store it for that key. Otherwise, we store 1-element list of that modifier for that key

Differential Revision: D50173854

fbshipit-source-id: e3da9fe96b3fd008219a8cdf00c7103e10c8f43f