buck2_error: Fix `#[from]`
Summary:
We need to forward the impls that are added to the inner type to the outer one

Stiopa is going to say that this kind of thing is proof that the derive macro is too complex. He's not wrong. But I think making it just work for now is the right thing to do, and in the future we should probably move away from using derive macros like this for error definitions

Reviewed By: stepancheg

Differential Revision: D50934084

fbshipit-source-id: 1984347a81c35efb17967dd87058fc9da46b4538