buck2_error: More powerful downcast
Summary:
This fixes a downcast bug. Mostly explained in the diff.

This is all pretty ugly. The comment does include a suggestion for how this uglyness might be mitigated, but that would require a much bigger change which is not clearly easier to maintain and would only mitigate some of the uglyness.

The proper fix for this is to just stop downcasting things.

Reviewed By: stepancheg

Differential Revision: D50959021

fbshipit-source-id: 46ec259fbed9f62a6f89abd28e43c9b68120bc44