Return both built targets and errors always
Summary:
There was a decision made to return a result of good/bad that was either the built targets or the errors if anything failed.  This worked mostly file since the INPUT target patterns is what is logged internally.
But up the stack, we want to track the rule_names of the built targets.  In which case, on a multi pattern or a request to build A & B, we would want to know the rules for each built target even in the case of one target failing.

Reviewed By: JakobDegen

Differential Revision: D50699175

fbshipit-source-id: 1e48bed1df2e3ebab5ca660a1b9ede8d2847b013