Do not merge library manifests into module AndroidManifest.xml#
Summary:
In buck1 (and the expectation) of Voltron today is that we do not merge
module manifests into the app module's AndroidManifest.xml

This is because we do not always install the module via the
PackageManager. This owuld cause an issue if the code that is running
expects things to exist.

Reviewed By: IanChilds

Differential Revision: D50977242

fbshipit-source-id: c10ed3458d8c9fd80aa5cc584a0eecdc88704f74