Add PACKAGE modifiers to cfg constructor
Summary: Hooks up PACKAGE modifiers with cfg constructor function, completing the core side of integrating PACKAGE modifiers. This diff does it by putting `TargetLabel` on the dice key and then getting PACKAGE modifiers from the `TargetNode`. The PACKAGE modifiers are allocate on Starlark heap as a dictionary and then passed to the pre constraint analysis function as `package_modifiers`.

Reviewed By: stepancheg

Differential Revision: D50673909

fbshipit-source-id: 0f2e1ff8cd4d0ae2d49a9330df68a48687aa1a00