buck2_error: Greatly simplify late formatting machinery
Summary:
I was high the first time I wrote this (/s). This time I'm not.

Putting the late formatter on the error root is sort of wrong anyway - if we have an HTTP error, we may not know how we want to format it when it's shown at the end of the build until it's actually emitted for the first time and we know what to do with the error.

It's also unnecessarily complicated. We can just store the late formatter when we emit the error, which results in much less - and far more straightforward - code.

There's still more cleanup possible here, but this diff is big enough as is.

Reviewed By: stepancheg

Differential Revision: D51068382

fbshipit-source-id: 27223db8f8439e7cab81321cfa24437274cc4222