rust-project: discover buck targets on the fly
Summary: This diff adds support for automatic buck target indexing in rust-analyzer. The changes here (in the rust-companion extension and `rust-project`) forward a notification from rust-analyzer that the file itself is not indexed and report on *which* target is being indexed to the user. Releasing this change is dependent on the changes in https://github.com/davidbarsky/rust-analyzer/tree/davidbarsky/start-of-monorepo-mode, which I will open a separate diff for.

Reviewed By: Wilfred

Differential Revision: D51099597

fbshipit-source-id: f7ff32f80a04c3f9b71d0ff70a93f67956a7618f