errors: Flip action error creation order
Summary:
Before this diff, errors were created in the following order `ExecuteError -> buck2_data::ActionError -> buck2_build_api::ActionError`. This is a problem, because it means that error metadata gets lost when we serialize things to the protobuf event.

This diff flips the last two steps around, so that it's `ExecuteError -> buck2_build_api::ActionError -> buck2_data::ActionError`. That unblocks the next few diffs in which we can actually retain the action error's metadata

Reviewed By: stepancheg

Differential Revision: D51110453

fbshipit-source-id: f40bd6d61909d542c3f393b5544727bb13f67091