errors: Make `provide_metadata` public and documented
Summary:
Explicitly endorse that people can use this manually if they want to.

Using this is unfortunately pretty complicated, and ideally the derive macro would be good enough to cover all cases that we need to. Unfortunately, the case that I run into in the next diff is not well-supported by the derive macro. The specific issue is with manual `Display` impls. Due to compiler implementations (hopefully fixed soon), they run into issues with apparently (but not really) overlapping impls.

The alternative to this diff would be to either add an explicit `no_display` option to the derive macro or attempt to duplicate `thiserror`'s logic to infer whether the impl should be generated. Neither of those options really seem better.

Reviewed By: stepancheg

Differential Revision: D51110480

fbshipit-source-id: 61681e14da5c36346ac279fe65584823953cdb3b